Added write ability, removed ramfs
Added the routines to handle writing to files.  It's currently extremely
inefficient, and has some nasty design issues that could result in the
entire file disappearing if the user cancels the write at the wrong
time.  These will be fixed in the future.

Also removed the example ramfs code, as it wasn't being used at all.